FullyQualifiedGlobal{Constants,Functions}Sniff: Refactor logic into shared ancestor
