+++ SlevomatCodingStandard/Sniffs/Namespaces/FullyQualifiedGlobalFunctionsSniff.php
@@ -2,92 +2,25 @@
 
 namespace SlevomatCodingStandard\Sniffs\Namespaces;
 
-use SlevomatCodingStandard\Helpers\NamespaceHelper;
-use SlevomatCodingStandard\Helpers\ReferencedNameHelper;
-use SlevomatCodingStandard\Helpers\SniffSettingsHelper;
-use SlevomatCodingStandard\Helpers\UseStatementHelper;
+use SlevomatCodingStandard\Helpers\ReferencedName;
 
-class FullyQualifiedGlobalFunctionsSniff implements \PHP_CodeSniffer\Sniffs\Sniff
+class FullyQualifiedGlobalFunctionsSniff
+	extends \SlevomatCodingStandard\Sniffs\Namespaces\AbstractFullyQualifiedGlobalReference
 {
 
-	public const CODE_NON_FULLY_QUALIFIED = 'NonFullyQualified';
-
-	/** @var string[] */
-	public $exclude = [];
-
-	/** @var string[]|null */
-	private $normalizedExclude;
-
-	/**
-	 * @return mixed[]
-	 */
-	public function register(): array
+	protected function getNotFullyQualifiedMessage(): string
 	{
-		return [
-			T_OPEN_TAG,
-		];
+		return 'Function %s() should be referenced via a fully qualified name.';
 	}
 
-	/**
-	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
-	 * @param \PHP_CodeSniffer\Files\File $phpcsFile
-	 * @param int $openTagPointer
-	 */
-	public function process(\PHP_CodeSniffer\Files\File $phpcsFile, $openTagPointer): void
+	protected function isCaseSensitive(): bool
 	{
-		$tokens = $phpcsFile->getTokens();
-
-		$referencedNames = ReferencedNameHelper::getAllReferencedNames($phpcsFile, $openTagPointer);
-		$useStatements = UseStatementHelper::getUseStatements($phpcsFile, $openTagPointer);
-		$exclude = array_flip($this->getNormalizedExclude());
-
-		foreach ($referencedNames as $referencedName) {
-			$name = $referencedName->getNameAsReferencedInFile();
-			$canonicalName = strtolower($name);
-			$referenceNamePointer = $referencedName->getStartPointer();
-
-			if (!$referencedName->isFunction()) {
-				continue;
-			}
-
-			if (NamespaceHelper::isFullyQualifiedName($name)) {
-				continue;
-			}
-
-			if (NamespaceHelper::hasNamespace($name)) {
-				continue;
-			}
-
-			if (array_key_exists($canonicalName, $useStatements)) {
-				continue;
-			}
-
-			if (array_key_exists($canonicalName, $exclude)) {
-				continue;
-			}
-
-			$fix = $phpcsFile->addFixableError(sprintf('Function %s() should be referenced via a fully qualified name.', $tokens[$referenceNamePointer]['content']), $referenceNamePointer, self::CODE_NON_FULLY_QUALIFIED);
-			if ($fix) {
-				$phpcsFile->fixer->beginChangeset();
-				$phpcsFile->fixer->addContentBefore($referenceNamePointer, NamespaceHelper::NAMESPACE_SEPARATOR);
-				$phpcsFile->fixer->endChangeset();
-			}
-		}
+		return true;
 	}
 
-	/**
-	 * @return string[]
-	 */
-	private function getNormalizedExclude(): array
+	protected function isValidType(ReferencedName $name): bool
 	{
-		if ($this->normalizedExclude === null) {
-			$exclude = array_map(function (string $name): string {
-				return strtolower($name);
-			}, SniffSettingsHelper::normalizeArray($this->exclude));
-
-			$this->normalizedExclude = $exclude;
-		}
-		return $this->normalizedExclude;
+		return $name->isFunction();
 	}
 
 }
